pkg: move repo-related functions to pkg/auth
Move GetIndexName(), SplitReposName() to pkg/auth, to be consistent
with locations of other functions.
