+++ pkg/auth/auth.go
@@ -32,6 +32,10 @@ import (
 var (
 	defaultScopeAccess        = "pull"
 	defaultRegURL      string = "docker.io/busybox:latest"
+
+	DefaultIndexURLPlain = "registry-1.docker.io"
+	DefaultIndexURLAuth  = "index.docker.io"
+	DefaultRepoPrefix    = "library/"
 )
 
 type AuthScope struct {
@@ -258,4 +262,24 @@ func parseScope(inputScope string) AuthScope {
 	}
 
 	return outScope
+}
+
+// GetIndexName returns the index server from a registry URL.
+func GetIndexName(regURL string) string {
+	index, _ := SplitReposName(regURL)
+	return index
+}
+
+// SplitReposName breaks a repo name into an index name and remote name.
+func SplitReposName(name string) (indexName, remoteName string) {
+	i := strings.IndexRune(name, '/')
+	if i == -1 || (!strings.ContainsAny(name[:i], ".:") && name[:i] != "localhost") {
+		indexName, remoteName = DefaultIndexURLPlain, name
+	} else {
+		indexName, remoteName = name[:i], name[i+1:]
+	}
+	if indexName == DefaultIndexURLPlain && !strings.ContainsRune(remoteName, '/') {
+		remoteName = DefaultRepoPrefix + remoteName
+	}
+	return
 }
