+++ pkg/image/util.go
@@ -13,33 +13,3 @@
 // limitations under the License.
 
 package image
-
-import (
-	"strings"
-)
-
-const (
-	DefaultIndexURLPlain = "registry-1.docker.io"
-	DefaultIndexURLAuth  = "index.docker.io"
-	DefaultRepoPrefix    = "library/"
-)
-
-// GetIndexName returns the index server from a registry URL.
-func GetIndexName(regURL string) string {
-	index, _ := SplitReposName(regURL)
-	return index
-}
-
-// SplitReposName breaks a repo name into an index name and remote name.
-func SplitReposName(name string) (indexName, remoteName string) {
-	i := strings.IndexRune(name, '/')
-	if i == -1 || (!strings.ContainsAny(name[:i], ".:") && name[:i] != "localhost") {
-		indexName, remoteName = DefaultIndexURLPlain, name
-	} else {
-		indexName, remoteName = name[:i], name[i+1:]
-	}
-	if indexName == DefaultIndexURLPlain && !strings.ContainsRune(remoteName, '/') {
-		remoteName = DefaultRepoPrefix + remoteName
-	}
-	return
-}
